split into ThrowSpeciticity + CatchSpecificity
